# docs/swagger.yaml
basePath: /
definitions:
  database.CostStatistics:
    properties:
      averageFee:
        type: number
      earliestSlot:
        type: integer
      latestSlot:
        type: integer
      maxFee:
        type: integer
      minFee:
        type: integer
      totalFees:
        type: integer
      totalTransactions:
        type: integer
    type: object
  database.GetAllObjectIDsResult:
    properties:
      limit:
        type: integer
      objectIds:
        items:
          type: string
        type: array
      offset:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Merkle Oracle Node API
  version: 1.0.0
paths:
  /healthcheck:
    get:
      description: Returns 200 if the service is alive.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Health Check
      tags:
      - system
  /objects:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all tracked object IDs from the database.
      parameters:
      - description: 'Number of items per page (default: 20, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of object IDs
          schema:
            $ref: '#/definitions/database.GetAllObjectIDsResult'
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Object IDs
      tags:
      - objects
  /objects/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an object by its ID from the database with key-value
        pairs. Optionally specify a timestamp to get historical state.
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      - description: RFC3339 timestamp to get object state at specific time (e.g.
          2024-01-15T10:30:00Z). If not provided, returns current state.
        in: query
        name: timestamp
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Object data with timestamp
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID format or timestamp
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Object not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Object by ID
      tags:
      - objects
  /objects/active:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of object IDs that have at least one
        non-deleted key from the database.
      parameters:
      - description: 'Number of items per page (default: 20, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of active object IDs
          schema:
            $ref: '#/definitions/database.GetAllObjectIDsResult'
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Active Object IDs
      tags:
      - objects
  /objects/deleted:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of object IDs that have no non-deleted
        keys from the database.
      parameters:
      - description: 'Number of items per page (default: 20, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of deleted object IDs
          schema:
            $ref: '#/definitions/database.GetAllObjectIDsResult'
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Deleted Object IDs
      tags:
      - objects
  /statistics/costs:
    get:
      consumes:
      - application/json
      description: Retrieves transaction fee statistics including total fees, averages,
        min/max values, and slot information.
      produces:
      - application/json
      responses:
        "200":
          description: Cost statistics
          schema:
            $ref: '#/definitions/database.CostStatistics'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Cost Statistics
      tags:
      - statistics
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Object Queries
  name: objects
- description: System status and health monitoring endpoints
  name: system
- description: Statistics and analytics endpoints
  name: statistics